change data type of attribute Supply.price to int

+++ src/backend/mariaDB/setupDatabase.sql
@@ -78,7 +78,7 @@ create table Supply (
    id int auto_increment,
    productId int not null,
    distributorId int not null,
-   price decimal(8, 2),
+   price int,
    buyDate Date not null,
    expirationDate Date not null,
    consumptionDate Date,
